Check all allocations for integer overflow
Calculations of the size (in nbytes) of an allocation can potentially
overflow. Sometimes because the size is calculated separately and the
calculation does not check for (int)a * int(b) > max_int and similar
for other data types.

The other main case is the calculation being performed actually in the
call to the allocation function.

Try to check these all. Use a simple function to check for overflow,
returns the result and 0 on success, or -1 on failure.

In general the calculations performed in the call use ints, so just cast
at least one variable to a size_t to promote the variables to the
data type used by the allocation routines. This does not check for
exceeding max_size_t.

+++ base/gdevprn.c
@@ -1544,7 +1544,7 @@ gx_default_setup_buf_device(gx_device *bdev, byte *buffer, int raster,
         ptrs = (byte **)
             gs_alloc_byte_array(mdev->memory,
                                 (mdev->num_planar_planes ?
-                                 full_height * mdev->num_planar_planes :
+                                 (size_t)full_height * mdev->num_planar_planes :
                                  setup_height),
                                 sizeof(byte *), "setup_buf_device");
         if (ptrs == 0)
